requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
pdf2image>=1.16.3
pillow>=10.0.0
python-multipart>=0.0.6
rich>=13.0.0
pydantic>=2.0.0
